refactor: demo page connect wallet modal props fulfillment as a component
