+++ apps/demo-react/components/connect-wallet-modal/index.ts
@@ -1,0 +1,1 @@
+export * from './connect-wallet-modal';
